Fix rewriting not going deep enough.
In some places nodes were being returned directly
or just having part of their nodes replaced.
This is an issue because we have up to trivia rewriting and we need to go down to the token level for this to have effect,
and the easiest way to do so is to call the base class' visit method
instead of reimplementing it ourselves.